split context token from implementation
