+++ src/packages/documents/documents/repository/detail/document-detail.store.ts
@@ -1,7 +1,7 @@
-import type { UmbDocumentDetailModel } from '../../types.js';
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 import { UmbDetailStoreBase } from '@umbraco-cms/backoffice/store';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
+import type { UmbDocumentDetailModel } from '../../types.js';
+import { UMB_DOCUMENT_DETAIL_STORE_CONTEXT } from './document-detail.store.context-token.js';
 
 /**
  * @export
@@ -21,5 +21,3 @@ export class UmbDocumentDetailStore extends UmbDetailStoreBase<UmbDocumentDetail
 }
 
 export { UmbDocumentDetailStore as api };
-
-export const UMB_DOCUMENT_DETAIL_STORE_CONTEXT = new UmbContextToken<UmbDocumentDetailStore>('UmbDocumentDetailStore');
