+++ src/packages/documents/documents/repository/detail/index.ts
@@ -1,3 +1,3 @@
 export { UmbDocumentDetailRepository } from './document-detail.repository.js';
 export { UMB_DOCUMENT_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
-export { UMB_DOCUMENT_DETAIL_STORE_CONTEXT } from './document-detail.store.js';
+export { UMB_DOCUMENT_DETAIL_STORE_CONTEXT } from './document-detail.store.context-token.js';
